Separate callHook from hooks.
This prevents scripts from crashing when they're not compiled
by webpack (i.e. gluestick auto-upgrade).
